requirements.txt
mpl_finance==0.10.1
requests~=2.31.0
pandas~=2.2.2
beautifulsoup4~=4.12.3
nsepy~=0.8
SQLAlchemy~=1.4.39
nsetools~=1.0.11
python-dateutil~=2.8.2
numpy~=1.24.1
yfinance~=0.2.12
matplotlib~=3.6.3
plotly~=5.16.1
pillow~=10.3.0
fuzzywuzzy~=0.18.0
pytz~=2023.3.post1
mplfinance~=0.12.10b0
statsmodels~=0.14.0
xlwings~=0.27.15
QuantLib==1.28
scipy~=1.10.0
nltk~=3.7
stockstats~=0.4.1
keras~=2.9.0
Django~=4.2.5
PyQt5~=5.15.10
py2exe~=0.13.0.0
setuptools~=68.2.2
tkcalendar~=1.6.1
kiteconnect~=5.0.1
sympy~=1.12.1
pandas~=1.5.2